Code refactoring for new GNNs of schnet, forcenet, dimenet++. Small log changes
