+++ configs/is2re/100k/schnet/new_schnet.yml
@@ -1,9 +1,10 @@
 includes:
-- configs/is2re/all/base.yml
+- configs/is2re/100k/base.yml
 
 model:
-  name: schnet
+  name: new_schnet
   hidden_channels: 384
+  tag_hidden_channels: 64
   num_filters: 128
   num_interactions: 4
   num_gaussians: 100
@@ -12,20 +13,20 @@ model:
   regress_forces: False
 
 # *** Important note ***
-#   The total number of gpus used for this run was 4.
+#   The total number of gpus used for this run was 1.
 #   If the global batch size (num_gpus * batch_size) is modified
 #   the lr_milestones and warmup_steps need to be adjusted accordingly.
 
 optim:
-  batch_size: 64
-  eval_batch_size: 64
+  batch_size: 32
+  eval_batch_size: 32
   num_workers: 16
-  lr_initial: 0.001
+  lr_initial: 0.0005
   lr_gamma: 0.1
   lr_milestones: # steps at which lr_initial <- lr_initial * lr_gamma
-    - 17981
-    - 26972
-    - 35963
-  warmup_steps: 5394
+    - 15625
+    - 31250
+    - 46875
+  warmup_steps: 9375
   warmup_factor: 0.2
   max_epochs: 30
