+++ configs/is2re/10k/schnet/new_schnet.yml
@@ -1,18 +1,19 @@
 includes:
-- configs/is2re/all/base.yml
+- configs/is2re/10k/base.yml
 
 model:
-  name: schnet
-  hidden_channels: 384
+  name: new_schnet
+  hidden_channels: 256
+  tag_hidden_channels: 32
   num_filters: 128
-  num_interactions: 4
+  num_interactions: 3
   num_gaussians: 100
   cutoff: 6.0
   use_pbc: True
   regress_forces: False
 
 # *** Important note ***
-#   The total number of gpus used for this run was 4.
+#   The total number of gpus used for this run was 1.
 #   If the global batch size (num_gpus * batch_size) is modified
 #   the lr_milestones and warmup_steps need to be adjusted accordingly.
 
@@ -20,12 +21,12 @@ optim:
   batch_size: 64
   eval_batch_size: 64
   num_workers: 16
-  lr_initial: 0.001
+  lr_initial: 0.005
   lr_gamma: 0.1
-  lr_milestones: # steps at which lr_initial <- lr_initial * lr_gamma
-    - 17981
-    - 26972
-    - 35963
-  warmup_steps: 5394
+  lr_milestones: # epochs at which lr_initial <- lr_initial * lr_gamma
+    - 1562
+    - 2343
+    - 3125
+  warmup_steps: 468
   warmup_factor: 0.2
   max_epochs: 30
