+++ configs/is2re/all/dimenet_plus_plus/new_dpp.yml
@@ -2,14 +2,18 @@ includes:
 - configs/is2re/all/base.yml
 
 model:
-  name: schnet
-  hidden_channels: 384
-  num_filters: 128
-  num_interactions: 4
-  num_gaussians: 100
+  name: new_dimenetplusplus
+  hidden_channels: 256
+  out_emb_channels: 192
+  num_blocks: 3
   cutoff: 6.0
-  use_pbc: True
+  num_radial: 6
+  num_spherical: 7
+  num_before_skip: 1
+  num_after_skip: 2
+  num_output_layers: 3
   regress_forces: False
+  use_pbc: True
 
 # *** Important note ***
 #   The total number of gpus used for this run was 4.
@@ -17,15 +21,15 @@ model:
 #   the lr_milestones and warmup_steps need to be adjusted accordingly.
 
 optim:
-  batch_size: 64
-  eval_batch_size: 64
-  num_workers: 16
-  lr_initial: 0.001
+  batch_size: 4
+  eval_batch_size: 4
+  num_workers: 4
+  lr_initial: 0.0001
   lr_gamma: 0.1
   lr_milestones: # steps at which lr_initial <- lr_initial * lr_gamma
-    - 17981
-    - 26972
-    - 35963
-  warmup_steps: 5394
+    - 115082
+    - 230164
+    - 345246
+  warmup_steps: 57541
   warmup_factor: 0.2
-  max_epochs: 30
+  max_epochs: 20
