+++ configs/is2re/all/forcenet/new_forcenet.yml
@@ -1,0 +1,37 @@
+includes:
+- configs/is2re/all/base.yml
+
+model:
+  name: new_forcenet
+  num_interactions: 5
+  cutoff: 6
+  basis: "sphallmul"
+  ablation: "none"
+  depth_mlp_edge: 2
+  depth_mlp_node: 1
+  activation_str: "swish"
+  decoder_activation_str: "swish"
+  feat: "full"
+  hidden_channels: 512
+  decoder_hidden_channels: 512
+  max_n: 3
+
+# *** Important note ***
+#   The total number of gpus used for this run was 8.
+#   If the global batch size (num_gpus * batch_size) is modified
+#   the lr_milestones and warmup_steps need to be adjusted accordingly.
+
+optim:
+  batch_size: 8
+  eval_batch_size: 8
+  num_workers: 8
+  lr_initial: 0.0005
+  max_epochs: 20
+  energy_coefficient: 0
+  lr_gamma: 0.1
+  lr_milestones: # steps at which lr_initial <- lr_initial * lr_gamma
+    - 15625
+    - 25000
+    - 31250
+  warmup_steps: 9375
+  warmup_factor: 0.2
